# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kayak rental platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin — list orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: date_from
          schema:
            type: ["string", "null"]
        - in: query
          name: date_to
          schema:
            type: ["string", "null"]
        - in: query
          name: search
          schema:
            type: ["string", "null"]
        - in: query
          name: page
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrdersResponse"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin]
      summary: Admin — get order detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin — update order status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
...
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin — list all products (including inactive)
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProduct"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Admin — create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin — update product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminOrdersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrder"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer
      required:
        - data
        - total
        - page
        - perPage
    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryType:
          type: string
        exactPrice:
          type: ["number", "null"]
        approximatePrice:
          type: ["number", "null"]
        managerReviewRequired:
          type: boolean
        createdAt:
          type: string
        itemCount:
          type: integer
      required:
        - id
        - orderNumber
        - status
        - startDate
        - endDate
        - customerName
        - customerPhone
        - deliveryType
        - managerReviewRequired
        - createdAt
        - itemCount
    AdminOrderDetail:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        communicationChannel:
          type: string
        deliveryType:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        exactPrice:
          type: ["number", "null"]
        approximatePrice:
          type: ["number", "null"]
        managerReviewRequired:
          type: boolean
        pricingComment:
          type: ["string", "null"]
        totalDeposit:
          type: ["number", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusHistory"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - startDate
        - endDate
        - customerName
        - customerPhone
        - communicationChannel
        - deliveryType
        - managerReviewRequired
        - items
        - statusHistory
        - createdAt
        - updatedAt
    AdminOrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        lifejacketSizes:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]
      required:
        - id
        - productId
        - productName
        - quantity
    OrderStatusHistory:
      type: object
      properties:
        status:
          type: string
        comment:
          type: ["string", "null"]
        changedAt:
          type: string
      required:
        - status
        - changedAt
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [new, confirmed, paid, assembled, issued, delivered, completed, cancelled]
        comment:
          type: ["string", "null"]
      required:
        - status
...
    AdminProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        active:
          type: boolean
        featured:
          type: boolean
        depositAmount:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        constructionType:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - categoryId
        - categoryName
        - active
        - featured
        - createdAt
    AdminProductInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: integer
        shortDescription:
          type: ["string", "null"]
        fullDescription:
          type: ["string", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        depositAmount:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        constructionType:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        kit:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        h1:
          type: ["string", "null"]
      required:
        - name
        - slug
        - categoryId
        - active
        - featured
...